GUVNOR-2273 : Replace "package name white list" with DependencyFilter and related UI improvements
