+++ guvnor-services/guvnor-services-backend/src/main/java/org/guvnor/common/services/backend/config/SafeSessionInfo.java
@@ -19,7 +19,9 @@ import org.jboss.errai.security.shared.api.identity.User;
 import org.jboss.errai.security.shared.api.identity.UserImpl;
 import org.uberfire.rpc.SessionInfo;
 
-public class SafeSessionInfo implements SessionInfo {
+
+public class SafeSessionInfo
+        implements SessionInfo {
 
     private SessionInfo delegate;
