+++ guvnor-webapp/src/main/java/org/guvnor/server/MavenProjectServiceImpl.java
@@ -16,6 +16,7 @@
 
 package org.guvnor.server;
 
+import java.util.Set;
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.event.Event;
 import javax.inject.Inject;
@@ -23,11 +24,10 @@ import javax.inject.Named;
 
 import org.apache.commons.lang3.StringUtils;
 import org.guvnor.common.services.backend.exceptions.ExceptionUtilities;
+import org.guvnor.common.services.backend.util.CommentedOptionFactory;
 import org.guvnor.common.services.project.backend.server.AbstractProjectService;
-import org.guvnor.common.services.project.backend.server.ProjectConfigurationContentHandler;
 import org.guvnor.common.services.project.backend.server.utils.IdentifierUtils;
 import org.guvnor.common.services.project.builder.events.InvalidateDMOProjectCacheEvent;
-import org.guvnor.common.services.project.events.DeleteProjectEvent;
 import org.guvnor.common.services.project.events.NewPackageEvent;
 import org.guvnor.common.services.project.events.NewProjectEvent;
 import org.guvnor.common.services.project.events.RenameProjectEvent;
@@ -36,6 +36,7 @@ import org.guvnor.common.services.project.model.Package;
 import org.guvnor.common.services.project.model.Project;
 import org.guvnor.common.services.project.service.POMService;
 import org.guvnor.common.services.project.service.ProjectService;
+import org.guvnor.structure.backend.backcompat.BackwardCompatibleUtil;
 import org.guvnor.structure.server.config.ConfigurationFactory;
 import org.guvnor.structure.server.config.ConfigurationService;
 import org.jboss.errai.bus.server.annotations.Service;
@@ -43,76 +44,49 @@ import org.uberfire.backend.server.util.Paths;
 import org.uberfire.backend.vfs.Path;
 import org.uberfire.io.IOService;
 import org.uberfire.java.nio.file.FileSystem;
-import org.uberfire.java.nio.file.Files;
 import org.uberfire.rpc.SessionInfo;
+import org.uberfire.security.authz.AuthorizationManager;
 
-import org.jboss.errai.security.shared.api.identity.User;
+import static org.guvnor.common.services.project.backend.server.ProjectResourcePaths.*;
 
 
 @Service
 @ApplicationScoped
-public class ProjectServiceImpl
+public class MavenProjectServiceImpl
         extends AbstractProjectService<Project>
         implements ProjectService<Project> {
 
-    public ProjectServiceImpl() {
+    public MavenProjectServiceImpl() {
         // Boilerplate sacrifice for Weld
     }
 
     @Inject
-    public ProjectServiceImpl( @Named("ioStrategy") IOService ioService,
-                               POMService pomService,
-                               ProjectConfigurationContentHandler projectConfigurationContentHandler,
-                               ConfigurationService configurationService,
-                               ConfigurationFactory configurationFactory,
-                               Event<NewProjectEvent> newProjectEvent,
-                               Event<NewPackageEvent> newPackageEvent,
-                               Event<RenameProjectEvent> renameProjectEvent,
-                               Event<DeleteProjectEvent> deleteProjectEvent,
-                               Event<InvalidateDMOProjectCacheEvent> invalidateDMOCache,
-                               User identity,
-                               SessionInfo sessionInfo ) {
-        super( ioService, pomService, projectConfigurationContentHandler, configurationService,
-               configurationFactory, newProjectEvent, newPackageEvent, renameProjectEvent, deleteProjectEvent,
-               invalidateDMOCache, identity, sessionInfo );
-    }
-
-    @Override
-    public Project resolveProject( final Path resource ) {
-        try {
-            //Null resource paths cannot resolve to a Project
-            if ( resource == null ) {
-                return null;
-            }
-
-            //Check if resource is the project root
-            org.uberfire.java.nio.file.Path path = Paths.convert( resource ).normalize();
-
-            //A project root is the folder containing the pom.xml file. This will be the parent of the "src" folder
-            if ( Files.isRegularFile( path ) ) {
-                path = path.getParent();
-            }
-            if ( hasPom( path ) ) {
-                return makeProject( path );
-            }
-            while ( path.getNameCount() > 0 && !path.getFileName().toString().equals( SOURCE_FILENAME ) ) {
-                path = path.getParent();
-            }
-            if ( path.getNameCount() == 0 ) {
-                return null;
-            }
-            path = path.getParent();
-            if ( path.getNameCount() == 0 || path == null ) {
-                return null;
-            }
-            if ( !hasPom( path ) ) {
-                return null;
-            }
-            return makeProject( path );
-
-        } catch ( Exception e ) {
-            throw ExceptionUtilities.handleException( e );
-        }
+    public MavenProjectServiceImpl( final @Named( "ioStrategy" ) IOService ioService,
+                                    final POMService pomService,
+                                    final ConfigurationService configurationService,
+                                    final ConfigurationFactory configurationFactory,
+                                    final Event<NewProjectEvent> newProjectEvent,
+                                    final Event<NewPackageEvent> newPackageEvent,
+                                    final Event<RenameProjectEvent> renameProjectEvent,
+                                    final Event<InvalidateDMOProjectCacheEvent> invalidateDMOCache,
+                                    final SessionInfo sessionInfo,
+                                    final AuthorizationManager authorizationManager,
+                                    final BackwardCompatibleUtil backward,
+                                    final CommentedOptionFactory commentedOptionFactory,
+                                    final MavenResourceResolver resourceResolver ) {
+        super( ioService,
+               pomService,
+               configurationService,
+               configurationFactory,
+               newProjectEvent,
+               newPackageEvent,
+               renameProjectEvent,
+               invalidateDMOCache,
+               sessionInfo,
+               authorizationManager,
+               backward,
+               commentedOptionFactory,
+               resourceResolver );
     }
 
     @Override
@@ -125,7 +99,8 @@ public class ProjectServiceImpl
             final Path fsRoot = repository.getRoot();
             final Path projectRootPath = Paths.convert( Paths.convert( fsRoot ).resolve( pom.getName() ) );
 
-            ioService.startBatch( new FileSystem[]{fs}, makeCommentedOption( "New project [" + pom.getName() + "]" ) );
+            ioService.startBatch( new FileSystem[]{fs},
+                                  commentedOptionFactory.makeCommentedOption( "New project [" + pom.getName() + "]" ) );
 
             //Create POM.xml
             pomService.create( projectRootPath,
@@ -134,7 +109,9 @@ public class ProjectServiceImpl
 
             //Raise an event for the new project
             final Project project = resolveProject( projectRootPath );
-            newProjectEvent.fire( new NewProjectEvent( project, getSessionId(), getIdentityName() ) );
+            newProjectEvent.fire( new NewProjectEvent( project,
+                                                       commentedOptionFactory.getSafeSessionId(),
+                                                       commentedOptionFactory.getSafeIdentityName() ) );
 
             //Create a default workspace based on the GAV
             final String legalJavaGroupId[] = IdentifierUtils.convertMavenIdentifierToJavaIdentifier( pom.getGav().getGroupId().split( "\\.",
@@ -146,9 +123,9 @@ public class ProjectServiceImpl
                                                                                                   "/" );
             final Path defaultPackagePath = Paths.convert( Paths.convert( projectRootPath ).resolve( MAIN_RESOURCES_PATH ) );
             final org.guvnor.common.services.project.model.Package defaultPackage = resolvePackage( defaultPackagePath );
-            final Package defaultWorkspacePackage = doNewPackage( defaultPackage,
-                                                                  defaultWorkspacePath,
-                                                                  false );
+            final Package defaultWorkspacePackage = resourceResolver.newPackage( defaultPackage,
+                                                                                 defaultWorkspacePath,
+                                                                                 false );
 
             //Raise an event for the new project's default workspace
             newPackageEvent.fire( new NewPackageEvent( defaultWorkspacePackage ) );
@@ -164,22 +141,52 @@ public class ProjectServiceImpl
     }
 
     @Override
-    public Project simpleProjectInstance( final org.uberfire.java.nio.file.Path nioProjectRootPath ) {
-        final Path projectRootPath = Paths.convert( nioProjectRootPath );
-
-        return new Project( projectRootPath,
-                            Paths.convert( nioProjectRootPath.resolve( POM_PATH ) ),
-                            projectRootPath.getFileName() );
-
+    public Project resolveProject( final Path resource ) {
+        return resourceResolver.resolveProject( resource );
     }
 
-  @Override
-  public Project resolveParentProject(Path resource) {
-    return null;
-  }
+    @Override
+    public Project resolveParentProject( final Path resource ) {
+        return resourceResolver.resolveParentProject( resource );
+    }
 
-  @Override
-  public Project resolveToParentProject(Path resource) {
-    return null;
-  }
+    @Override
+    public Project resolveToParentProject( final Path resource ) {
+        return resourceResolver.resolveParentProject( resource );
+    }
+
+    @Override
+    public Set<Package> resolvePackages( final Project project ) {
+        return resourceResolver.resolvePackages( project );
+    }
+
+    @Override
+    public Set<Package> resolvePackages( final Package pkg ) {
+        return resourceResolver.resolvePackages( pkg );
+    }
+
+    @Override
+    public Package resolveDefaultPackage( final Project project ) {
+        return resourceResolver.resolveDefaultPackage( project );
+    }
+
+    @Override
+    public Package resolveParentPackage( final Package pkg ) {
+        return resourceResolver.resolveParentPackage( pkg );
+    }
+
+    @Override
+    public boolean isPom( final Path resource ) {
+        return resourceResolver.isPom( resource );
+    }
+
+    @Override
+    public org.guvnor.common.services.project.model.Package resolvePackage( final Path resource ) {
+        return resourceResolver.resolvePackage( resource );
+    }
+
+    @Override
+    public Project simpleProjectInstance( final org.uberfire.java.nio.file.Path nioProjectRootPath ) {
+        return resourceResolver.simpleProjectInstance( nioProjectRootPath );
+    }
 }
